requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
astroid=2.5.6=pypi_0
autopep8=1.5.7=pypi_0
ca-certificates=2021.5.30=h5b45459_0
certifi=2021.5.30=py38haa244fe_0
chardet=4.0.0=pypi_0
colorama=0.4.4=pypi_0
docutils=0.17.1=py38haa244fe_0
flake8=3.9.2=pypi_0
freetype=2.10.4=h546665d_1
glew=2.1.0=h39d44d4_2
idna=2.10=pypi_0
isort=5.9.1=pypi_0
jbig=2.1=h8d14728_2003
jpeg=9d=h8ffe710_0
kivy=2.0.0=py38h7600b87_3
lazy-object-proxy=1.6.0=pypi_0
lcms2=2.12=h2a16943_0
lerc=2.2.1=h0e60522_0
libdeflate=1.7=h8ffe710_5
libpng=1.6.37=h1d00b33_2
libtiff=4.3.0=h0c97f57_1
lz4-c=1.9.3=h8ffe710_0
mccabe=0.6.1=pypi_0
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.4.0=hb211442_1
openssl=1.1.1k=h8ffe710_0
parso=0.8.2=pypi_0
pillow=8.3.1=py38h794f750_0
pip=21.1.2=py38haa95532_0
pluggy=0.13.1=pypi_0
pycodestyle=2.7.0=pypi_0
pydocstyle=6.1.1=pypi_0
pyflakes=2.3.1=pypi_0
pygments=2.9.0=pyhd8ed1ab_0
pylint=2.8.3=pypi_0
pyperclip=1.8.2=pypi_0
python=3.8.10=hdbf39b2_7
python-dateutil=2.8.1=pypi_0
python-lsp-jsonrpc=1.0.0=pypi_0
python-lsp-server=1.0.1=pypi_0
python_abi=3.8=2_cp38
requests=2.25.1=pypi_0
rope=0.19.0=pypi_0
sdl2=2.0.12=h0e60522_1
sdl2_image=2.0.5=h63225fd_0
sdl2_mixer=2.0.4=h6538335_0
sdl2_ttf=2.0.15=h4636d2b_0
setuptools=52.0.0=py38haa95532_0
six=1.16.0=pypi_0
smpeg2=2.0.0=h6538335_1
snowballstemmer=2.1.0=pypi_0
sqlite=3.35.4=h2bbff1b_0
tk=8.6.10=h8ffe710_1
toml=0.10.2=pypi_0
ujson=4.0.2=pypi_0
urllib3=1.26.6=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py38_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h62dcd97_1
yapf=0.31.0=pypi_0
zlib=1.2.11=h62dcd97_1010
zstd=1.5.0=h6255e5f_0
